pyproject.toml
[tool.poetry]
name = "NSscrapper"
version = "0.1.1"
description = "This is a scapper for UK NS eshop games"
authors = ["Zhining Li <lizhining0726@gmai.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.25.1"
urllib3 = "<1.26"
python-dateutil = "^2.9"
ddt = "^1.7.2"
algoliasearch = "^3.0.0"
bs4 = "^0.0.2" 

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
